Settings: Replace actionbar by toolbar:
As the detailed weather already uses a toolbar instead of an actionbar,
replace the actionbar by a toolbar in settings as well,
so we can use the (App)SplashTheme as a base theme for the app.
The weather settings can still be launched via the system settings (and DKSettings),
so we still need the "system" settings theme.

Also remove the AppSettingsSplashActivity and the SettingsSplashActivity.

TO-DO:
Fix settings theme when weather settings were launched via DKSettings,
and DKSettings was launched via launcher app drawér,
alternatively use the same (app) theme for DKSettings and DKWeather.